requirements.txt
asgiref==3.2.10
autopep8==1.5.3
Django==3.0.8
Pillow==7.2.0
pycodestyle==2.6.0
pytz==2020.1
sqlparse==0.3.1
toml==0.10.1
djangorestframework-recursive==0.1.2
django-mptt==0.14.0
psycopg2-binary==2.8.6